Fix IBV_WC_SEND not setting the byte_len to the transferred size
Since immediate mode data is not supported on BlueGene IO node netqwork roq,
we must send a short (4 byte) message instead of a 0 byte mesage and
this means we must also handle the completion as a buffer is used.

Since IBV_WC_SEND does not set the byte_len to the transferred
bytes on completion, we have a problem in checking if the send is a normal
one or a special tag send.

Currently, we assume that if the send does not match the active sends,
it must be a tag ack send of 4 bytes, but this assumption might cause trouble
one day.