If run using runtime_mode_connect, find the correct localhost ip address
If using --hpx::connect, the worker needs to know the public ip address
of itself so that it can send it to the hpx:agas root where it is used
for communication. If the user does not specify the hpx parcel address
than 127.0.0.1 is assumed, but this only works for localities on the
same node. This patch find the public ip address and uses that instead.

It may still fail if multiple ip addresses exist for the node,
but it is a better guess than before.

This is particularly usefult when the worker is spawned using
srun app <params>
and the user does not know which node sun will spwn the worker on
so cannot easily put the correct ip address in the command line.