Redesign connection code to improve support for N>2 nodes
We allow one thread to make a connection and others must wait
until the connection is complete before they continue.

The implementation is still not complete as an incoming connection
(from the same node we are connecting to) can still cause problems
if it happens during the ongoing connection, but this patch is the
first step towards a complete fix.