Fix a bug in the readers/writers lock
Readers increment the reader counter when acquiring the lock and
only increment writer count when releasing the lock. We were accidentally
incrementing the reader count on release as well as acquire.

Added a flag to allow us to handle a general 'unlock' whether we are
a shared lock or an exclusive one.

Added a couple of debug message options for next time this causes problems.