Improve custom scheduler stealing to better preserve cache coherency
when creating/adding tasks to queues - try to add them to the queue for
the current CPU/thread rather than using a round robin approach.

This helps work on one core to stay on that core.

Take tasks from front of own queue, but steal from back of others'
queues to make it more likely that work on own queue stays on own core.