Fix a barrier problem in network_storage test
If ranks 1,2,3...N enter or create a barrier before rank 0 has
initialized it, then their countdown is lost.
This patch adds a very simple barrier that is inefficient
but suitable for testing purposes and it does not have the race
problem as long as rank 0 inits the barrier before first use.
I can be used as many times as required after that.