Add new cross pool injection test
This test creates a random number of pools with a
random number of threads in each, it then creates
futures one one pool at random and continuations
on another with differnt combinations of priority
and some tasks waiting on others to test resumption.

It is intended as a stress test of scheduler management
of tasks/threads and resources etc.